feat(pvtx): create libptvx
This group of commit adds pvtx/CMakeLists.txt, which allows us to build
libpvtx.so, libpvtx.a, and of course pvtx and pvtx-static. It also enables us
to run tests without utilizing the main CMakeLists.txt.
The CMakeLists.txt file in Pantavisor was streamlined; simply include
this new file while maintaining all of its features.

feat(pvtx): improve help message

feat(pvtx): improve error reporting on pvtx cli

fix(pvtx): send args without modification from cli to pv_pvtx_txn_begin

fix(pvtx): cli arguments capture

feat(pvtx): create pvtx/CMakeLists.txt

This commit adds pvtx/CMakeLists.txt, which allows us to build
libpvtx_shared.so, libpvtx_static.a, and of course pvtx and pvtx-static.
It also enables us to run tests without utilizing the main
CMakeLists.txt.
The CMakeLists.txt file in Pantavisor was streamlined; simply include
this new file while maintaining all of its features.

refactor(pvtx): reorganize pvtx directory

fix(pvtx): make test works from pvtx and Pantavisor CMakeLists

feat(pvtx): rename library as libpvtx.so and libpvtx.a

feat(pvtx): tidy pvtx/CMakeLists

fix(pvtx): fix problems found with the new CMakeLists.txt

fix(pvtx): fix possible uninitialized use
