+++ pvtx/src/pvtx_error.c
@@ -1,4 +1,5 @@
-#include "pvtx_error.h"
+#include <pvtx/error.h>
+#include "utils/fs.h"
 
 #include <stdarg.h>
 #include <stdio.h>
